# BUG/PERF: MultiIndex setops with sort=None