--- a/asv_bench/benchmarks/multiindex_object.py
+++ b/asv_bench/benchmarks/multiindex_object.py
@@ -239,10 +239,11 @@ class SetOperations:
         ("monotonic", "non_monotonic"),
         ("datetime", "int", "string", "ea_int"),
         ("intersection", "union", "symmetric_difference"),
+        (False, None),
     ]
-    param_names = ["index_structure", "dtype", "method"]
+    param_names = ["index_structure", "dtype", "method", "sort"]
 
-    def setup(self, index_structure, dtype, method):
+    def setup(self, index_structure, dtype, method, sort):
         N = 10**5
         level1 = range(1000)
 
@@ -272,8 +273,8 @@ class SetOperations:
         self.left = data[dtype]["left"]
         self.right = data[dtype]["right"]
 
-    def time_operation(self, index_structure, dtype, method):
-        getattr(self.left, method)(self.right)
+    def time_operation(self, index_structure, dtype, method, sort):
+        getattr(self.left, method)(self.right, sort=sort)
 
 
 class Difference:
@@ -368,10 +369,14 @@ class Isin:
         }
 
         self.midx = data[dtype]
-        self.values = self.midx[:100]
+        self.values_small = self.midx[:100]
+        self.values_large = self.midx[:-1]
 
-    def time_isin(self, dtype):
-        self.midx.isin(self.values)
+    def time_isin_small(self, dtype):
+        self.midx.isin(self.values_small)
+
+    def time_isin_large(self, dtype):
+        self.midx.isin(self.values_large)
 
 
 from .pandas_vb_common import setup  # noqa: F401 isort:skip